Validar que usuario esté habilitado para editar

+++ src/main/java/com/portfolio/miportfolio/controller/ExperienciaRestController.java
@@ -3,8 +3,10 @@ package com.portfolio.miportfolio.controller;
 import com.portfolio.miportfolio.entity.Estudio;
 import com.portfolio.miportfolio.entity.Experiencia;
 import com.portfolio.miportfolio.service.IExperienciaService;
+import com.portfolio.miportfolio.service.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,6 +14,10 @@ import java.util.List;
 @RestController
 @RequestMapping("/api")
 public class ExperienciaRestController {
+
+    @Autowired
+    private UsuarioService usuarioService;
+
     @Autowired
     private IExperienciaService experienciaService;
 
@@ -36,6 +42,11 @@ public class ExperienciaRestController {
 
         Experiencia experienciaActual = experienciaService.findById(id);
 
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(experienciaActual.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
+
         experienciaActual.setEmpresa((experiencia.getEmpresa()));
         experienciaActual.setInicio(experiencia.getInicio());
         experienciaActual.setFin(experiencia.getFin());
@@ -50,6 +61,13 @@ public class ExperienciaRestController {
     @DeleteMapping("/experiencias/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
+
+        Experiencia experienciaAEliminar = experienciaService.findById(id);
+
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(experienciaAEliminar.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
         experienciaService.delete(id);
     }
